+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Next release
 
 - refactor(gateway): remove query-only support from user transactions
-- fix(gateway): serialize tip as hex in UserTransaction
+- fix(gateway): serialize `tip` as hex in UserTransaction
 - fix(block_production): dynamic block closing now adds special address with prev block hash
 - fix(rpc): call, simulate, estimate rpcs executed on top of the block, not at the start of it
 - fix(compilation): crate-level compilation
@@ -503,7 +503,6 @@
 
 ## v0.4.0
 
-- fix(gateway): serialize `tip` as hex in UserTransaction
 - chore: release v0.4.0
 - feat: better management of custom configurations for genesis assets
 - feat: use actual vm resource costs